requirements.txt
httpx==0.27.2
feedparser==6.0.11
tenacity==8.3.0
openai==1.55.3
jinja2==3.1.4
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pydantic==2.9.2
pytest==8.3.3
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
trafilatura==2.0.0
